Extract enums to separate files
