+++ Source/Extensions/Blazorise.Snackbar/Enums/SnackbarColor.cs
@@ -1,0 +1,52 @@
+﻿namespace Blazorise.Snackbar;
+
+/// <summary>
+/// Predefined set of contextual colors.
+/// </summary>
+public enum SnackbarColor
+{
+    /// <summary>
+    /// No color will be applied to an element.
+    /// </summary>
+    Default,
+
+    /// <summary>
+    /// Primary color.
+    /// </summary>
+    Primary,
+
+    /// <summary>
+    /// Secondary color.
+    /// </summary>
+    Secondary,
+
+    /// <summary>
+    /// Success color.
+    /// </summary>
+    Success,
+
+    /// <summary>
+    /// Danger color.
+    /// </summary>
+    Danger,
+
+    /// <summary>
+    /// Warning color.
+    /// </summary>
+    Warning,
+
+    /// <summary>
+    /// Info color.
+    /// </summary>
+    Info,
+
+    /// <summary>
+    /// Light color.
+    /// </summary>
+    Light,
+
+    /// <summary>
+    /// Dark color.
+    /// </summary>
+    Dark,
+}
